requirements.txt
Django==3.1.4
pymongo==3.11.3
pandas==1.2.1
dnspython==2.1.0
